pyproject.toml
[tool.poetry]
name = "signal-intelligence-backbone"
version = "0.1.0"
description = "Signal Intelligence Backbone for coherence tracking in multi-agent environments"
authors = ["Nithin Sarva <nithin@coherence.labs>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.0"
pydantic = "^2.5.0"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
streamlit = "^1.28.0"
plotly = "^5.17.0"
python-dotenv = "^1.0.0"
kafka-python = "^2.0.2"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
